requirements.txt
pandas
requests
duckdb
psycopg2-binary
python-dotenv
loguru

